[ALPS07792024] sapu: support FFA
[Detail]
replace 32-bit secure handle with 64-bit ffa handle

MTK-Commit-Id: 4423a4e4a7db982ec76689e54842ac73d6fe19ab

CR-Id: ALPS07792024
Feature: np_security

+++ drivers/misc/mediatek/apusys/sapu/sapu_driver.h
@@ -51,7 +51,7 @@ struct PWRarg {
 };
 
 struct haArg {
-	uint32_t handle;
+	uint64_t handle;
 	dma_addr_t dma_addr;
 	uint64_t mem_hd;
 };
